update: add test mode for run_epoch function.
