+++ test/bilstm-sequence-labelling.py
